refactor(patch): extract tests into dedicated module
